% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUIC.R
\name{matToNet}
\alias{matToNet}
\title{Inferring gene regulons from Gene Regulatory Network}
\usage{
matToNet(
  weightMat,
  methods = c("aracne", "clr", "mrnet", "mrnetb"),
  cutoff = NULL
)
}
\arguments{
\item{weightMat}{A square matrix whose element should be positive values.}

\item{methods}{The name of the network inference algorithm. Default: aracne.}

\item{cutoff}{Set the cutoff of regulatory networks. Default: NULL, means
inferring cutoff by itself.}
}
\value{
A data.frame or a list with regulatory networks.
}
\description{
Inferring gene regulons from gene regulatory network using weighted square matrix from PUIC output.
}
\examples{
data(expMat)
PUIC_res <- PUIC(expMat)
PUIC_net <- matToNet(PUIC_res)
}
