% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUIC.R
\name{PUIC}
\alias{PUIC}
\title{Inferring Gene Regulatory Network (GRN) using percentage of unique information contribution (PUIC) method}
\usage{
PUIC(
  expMat,
  regulators = NULL,
  targets = NULL,
  logScale = FALSE,
  ncores = 1,
  diag = c("auto", "zero", "one"),
  verbose = interactive()
)
}
\arguments{
\item{expMat}{A \code{matrix} storing the gene expression data. Rows corresponding
to features (eg. gene symbols) and columns corresponding to samples (cells).
Raw read counts, UMIs, TPMs or logNormalized counts were supported.}

\item{regulators}{The regulator genes used for GRN inferring (eg, transcription
factors). At least two regulators required. Default: NULL, using all the genes
in gene expression matrix. Default: NULL, all genes.}

\item{targets}{The target genes used for GRN inferring. Default: NULL, using
all the genes in gene expression matrix. Default: NULL, all genes.}

\item{logScale}{Whether to log-scale the input data? Default: FALSE.}

\item{ncores}{Number of cores used for parallel calculation. The running time
heavily depend on the number of regulators and targets, for genes over 5000,
we strongly suggest you to use multi-cores. Default: 1.}

\item{diag}{Numeric. The weight in the diagonal of output square matrix. Default: 1.}

\item{verbose}{Whether to print message while running. Default: TRUE.}
}
\value{
A matrix with weighted value between regulators and targets.
}
\description{
code used for calculating the \code{Multivariate
Information} (MI) and \code{specific.information}, discretizing gene expression data
were  partly adopted from package \code{\link{Informeasure}}.
}
\examples{
data(expMat)
PUIC_res <- PUIC(expMat)
head(PUIC_res)



}
